+++ src/ska_mid_dish_manager/component_managers/ds_cm.py
@@ -53,6 +53,7 @@ class DSComponentManager(TangoDeviceComponentManager):
             "dscpowerlimitkw",
             "trackTableCurrentIndex",
             "trackTableEndIndex",
+            "dscCtrlState",
         )
         super().__init__(
             tango_device_fqdn,